feat : Favorite 객체 생성
