--- src/main/java/jpa/domain/Favorite.java
@@ -1,29 +1,0 @@
-package jpa.domain;
-
-import lombok.*;
-import lombok.experimental.Accessors;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-@Builder
-@Accessors(chain = true)
-public class Favorite {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @CreatedDate
-    private LocalDateTime created_date;
-
-    @LastModifiedDate
-    private LocalDateTime modified_date;
-}
