+++ src/main/java/jpa/repository/FavoriteRepository.java
@@ -1,6 +1,6 @@
 package jpa.repository;
 
-import jpa.domain.Favorite;
+import jpa.entity.Favorite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
