+++ src/test/java/jpa/entity/FavoriteRepositoryTest.java
@@ -1,0 +1,15 @@
+package jpa.entity;
+
+import jpa.repository.FavoriteRepository;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@DataJpaTest
+public class FavoriteRepositoryTest {
+
+    @Autowired
+    FavoriteRepository favoriteRepository;
+}
